Add files via upload
Edited styling some. 
Added Quantify as secondary font.
Built media section at bottom of page for the gamelist, images of game selected, and a section for a description about game selected.(can be moved if we want).
Added login button that brings up window for user to log in. (need to implement google login into this section).